Moved `execute()` into samizdat-0.
I think the way things may end up is that there will just be a single
C codebase, that is not even have a separate sam-data library, though
the data-vs-not distinction will probably be maintained at least a bit
(to help keep the code clean). However, it wasn't looking like it was
way more trouble than it was worth to have a library(ish) barrier
between parsing and execution.