Implement sealer semantics for uniqlets.
I think this will help simplify some of the higher level
implementation, without really sacrificing the "purity" of the `dat`
module. Basically, this doesn't enable one to do anything that one
couldn't already do by storing uniqlets in a module-external mutable
table. And adding this stuff here makes it easier / simpler to do the
"impure" stuff that will inevitably need to be done as a matter of
practical programming in C.