I forgot to pass through the file open mode.
I also forgot to change `"r"` to `"w"` for the mode used to
open files for writing.

I also forgot to put a newline at the end of the silly
library string.

With this change, it looks like Samizdat Layer 0 can now
successfully produce output, trivial though it may be.

Yay!