Use `apply` to simplify `isStringlet()`.
Instead of manually destructuring the listlet, let the function
call mechanism do it for us. With just a little more rearranging,
it was also possible to remove the other inner function.

The old version of the code was the only client of `last()` and `butLast()`,
so remove those too. Maybe they'll come back.