Rearrange Samizdat Layer 0 core library.
This pulls out the code that was being prototyped in samizdat-1 along
with the (small amount of) code in `samizdat-0/lib/library.sam0`
to start a samizdat-0-lib module.

The code in samizdat-0-lib gets included in samizdat-0, but it
should also end up able to be compiled by samizdat-1. So, it made
sense to keep that code separate from both lower and upper
engines.

This change also exposes a bit more context functionality out
of the lang module in samizdat-0. I am pretty sure the way
contexts are used is going to have to change at some point, as
the current arrangement seems to be getting a bit unwieldy.