Make all the binary value comparators non-primitive.
Instead define them in terms of a newly-exposed `lowOrderIs`, which
does a comparison of two values and checks the result against one
or two possible values.

This also exposes `lowOrder` which is a direct reflection of the
underlying comparison, returning one of `-1 0 1`. However, note that
plain `lowOrder` won't suffice to write `eq` and friends (that is, you
need at least something more) in that you need at least one primitive
that can return a boolean result from a comparison.