Differentiate between `null` and `void`.
`null` is an in-model value (whose use will probably be pretty minimal
in the end). `void` (which doesn't really have a name in the language)
is the absence of a value, such as from a function that ends without
yielding (`<>`ing) a value.

`void` only ever appears as a yield/return value, where it is
acceptable to pass it along up a chain of yields, so long as
eventually that value (or rather non-value) gets dropped on the floor,
that is, as long as it isn't saved in a variable or passed as an
argument to some other function, and so on.

Fixes issue #11.