Rework how the core library gets exposed.
Instead of building up a single `zcontext` from the primitives and
then the `samizdat-0-lib` result, now `samizdat-0-lib` is in control
of what becomes the exported library. This means that, for example, it
can choose to omit some of the primitive definitions (which eventually
may be "unsafe" or "magic" and hence not appropriate to expose).

Added bonus: New maplet and context functions.