Rename and rearrange: `ifVoid` -> `ifValue`
It now parallels `ifTrue` in structure, taking either two or three
arguments, with second being the "if it got a value" one, and the
third being "if it got `void`" one. The second, if called, gets
called with the non-`void` value as an argument.

Partial fix (with not much left to go) for issue #23.