Make semicolon be a statement separator, not a statement terminator.
Note that with the rules as redefined here, there is no semicolon
required between the last statement of a function and the yield
diamond. This will probably want to change at some point, if we want
to have a concise ("wee") function form along the lines of what's
proposed in issue #21, as there would then be a bit of an ambiguity
between wee-function and function-yield.

Fixes issue #25.