Implement the Samizdat Layer 0 tokenizer in Samizdat Layer 0.
As a reasonably apples-to-apples comparison, this implementation
weighs in at about 190 lines, while the C version currently sits at
about 266 lines. This is a decent improvement in conciseness, but I
think things could be better.

Utility aside (utility will come, eventually), this was a great
learning exercise, exposing bugs, mis-features, and other enhancement
opportunities of the language.

Fixes issue #7.