Memorialize my understanding of fixpoint functions...
...fleeting though it may be. See the code comment for details, but
what I have here is a function in the "Y Combinator family" that
works for building self-recursive functions of arbitrary arity.

If Samizdat Layer 0 ends up "outlawing" use-before-def (which I think
it should), then something like this will need to be used to
set up recursive functions (at least if one wants to avoid the "hack"
of jumping through an `object`'s state, and one should so want that).

And if so, Samizdat Layer 0 will probably end up requiring this to
be used "manually", but a higher layer will end building these
sorts of functions automatically.