Rework the combinator to have two layers of function.
The point here is to make it so that the inner function doesn't
have to be modified at all in terms of its tree form, just needs
to have a different binding context.

I believe this form is trivially generalizable to a multi-function
mutual-recursion form, where you call the combinator with N functions,
and each wrapper has to be defined to take N arguments.