Switch tokenization model.
Now instead of raw single-character strings, the items when tokenizing
are in character-as-token form. That is, each element is a valueless
highlet whose type tag is the character to parse.

This paves the way for adding a useful data payload, such as source
position information, which could then be passed through to higher
layers.