Parenthesize all negative numbers.
This is to preemptively deal with the fact that `-N` will likely
become an expression (not just a constant when `N` is some digits),
and as such it won't be valid "naked" in an argument list or in a
list, map, or token literal.

Eventually, there may be a spacing-based way to make it be valid,
but that won't come for a while.