Give functions names.
This adds a `name` field to the low layer function implementation,
rather than just leaving it as a mapping in closure nodes. This
way, for example, primitive functions can (and now do) have names
that will get reported in stack traces, which is a big improvement
over having a forearm's length worth of `(unknown)`s.