Define a `voidable` expression type.
This makes it so that a function call argument that yields void
will only propagate voidness if the argument is marked "voidable".
If it isn't, it will now cause a loud and hard failure.

This change requires follow-on changes to (a) augment the grammar
so it is possible to mark arguments as voidable, and (b) use that
new syntax to mark up a bunch of places that need it.

(b) includes the expansion of comparison operators as well as
stuff written in samizdat-0 in a cheeky "avoid-and" way.