Implement a cache for `mapFind`.
On a hit, this avoids doing binary search to find a key in a map.
In practice, it looks like the theoretical maximum hit rate would
be something like 74% (that is, not quite 3 of 4 lookups are for
a map/key pair that have been looked up before). As configured,
the actual hit rate is about 73%, which seems to be the point of
diminishing returns in that increasing the cache size doesn't buy
much in the way of increased hit rate.