Expose `dataOf` `sizeOf` and `typeOf` as generics.
This involved a bit of re-plumbing to make functions and generics
both callable through a common C function. It'd be lovely if I
could get this to be all Ouroborosey -- a generic function for `call`
which dispatched to either Generic or Function -- but I'm not
entirely sure how to get there and if it's even advisable, other than
for some kind of idealistic sense of purity of essence, or something
like that.