Tweak the overflow definition for modulo.
For a standalone overflow-friendly math lib, it makes reasonable sense
to have div and mod have the same overflow characteristics, which is what
I did originally. However, in the context of being a "simulation" of
unbounded-size ints, it makes sense to have mod succeed when it *would*
succeed in a "realer" implementation.