Redefine how command-line running works.
It can now either run a file (as before) or run a directory. In the
latter case, the directory is taken to be a module definition, and this
mode is triggered by the existence of a `module.sam` file.

In the case of a file, instead of passing arguments, the global
environment now gets `ARGUMENTS` and `PATH` bindings. I expect this
will be more convenient in typical use cases. (Alas, it means I have
to go and adjust all the demos.)