Fix `moduleUse` search order.
This was a bit of a mess. The intention has always been that when
a module is loaded via `moduleUse`, the loaded module's `moduleUse`
can use sibling or descendant modules. In fact, the only thing that
ever worked was descendant modules of a top-level "main" module.

This change fixes things so that sibling modules can actually be found.
Based on some quick ad-hoc tests, this seems to work, but it needs a
little more testing before I'll call it totally done.