Get a little fancier about generating into an array.
In practice, I found I was blowing the stack because there were N
nested calls to collect a generator, with something like 100kbytes
of stack allocation per call. Now the stack allocation is much smaller,
and it falls back to a heap allocation in the rare cases where the
stack allocation is insufficient.