Tease out an `nthMapping` from `nth`.
This defines a new method `nthMapping` which replaces `nth` on maps,
but also works across all collections. `nth` is redefined to *only*
work on sequences.

This was a pretty messy change, because there were a lot of assumptions
about `nth` working on all collections. I couldn't just change everything
to `nthMapping` because that function always returns a mapping, and
it wouldn't be correct (for example) to use that when producing generated
values out of a list or string.

There's still some additional cleanup that will be needed.