Rework system startup.
Now the C code in `lib/init.c` is responsible for loading the module
system as a one-off, and using that to do a "proper" load on the library's
top-level `main.sam`. And `main.sam` is reworked to have a form factor much
more like an application's `main.sam`.

This change removes `boot.sam`, as everything it did has now been
absorbed by `init.c`. This makes the C code a little more complicated, but
makes the system as a whole just a bit simpler.