Specify and define `withoutTops`.
This separates out part of the functionality of `withSimpleDefs` into
its own function. `withSimpleDefs` will eventually have to be able
to do module resolution, but the functionality here (a) doesn't need
to do such resolution, and (b) needs to be called in a context where
it's not even appropriate to have resolution functionality.

This commit just adds the new function but doesn't remove the functionality
from `withSimpleDefs`, nor does it add any calls to this function.