Remove `interp-tree` mode.
With the addition of module resolution, there is no longer a
straightforward way to take text in from C code. More specifically,
it's more trouble than it's worth to set up a resolver function,
for a compiler mode that was only ever really meant as a first step,
and which has long since outlived its usefulness.