Change the spec for `parse()` regarding parsing failure.
Now, the spec says *not* to call `store()` at all on the given `box`
argument in the case of parsing failure. It used to specify calling
`store()` with no value argument. The reasons for the change are
(a) a void return already unambiguously signals parsing failure, and
(b) not touching the given `box` allows for reasonable reuse of a box
in many situations.