Be more frugal with value stack usage.
With this change, `stdCollect` now only grows the value stack by two
items per iteration, the yielded value and the sequence of generators.
Each generator *should* become garbage when the next generator is
returned, but that's unfortunately not easy to get the system to
realize. So, until I figure out a better solution, there's going to be
a bit more built-up garbage during `collect` than there ought to be;
at least, however, it should become freeable immediately after `collect`
returns.